worker/requirement.txt

google-cloud-storage
google-cloud-pubsub
google-cloud-speech
google-cloud-translate
functions-framework
ffmpeg-python
yt-dlp
requests
six
json5